refactor: decompose fetchAndMatch into staged methods
Extract the 500-line fetchAndMatch god function into 7 focused
methods in a new match.go file, threaded by a matchState struct:

  fetchRawListings        → fetch from source
  ingestCatalog           → update dynamic catalog
  matchListingsToSearches → percolator match + dedup + price detection
  publishEnrichRequests   → queue unenriched listings for enricher
  processPriceDrops       → format price drop notifications
  processAndDeliverPerSearch → pipeline + persist + grace window + deliver
  dropStaleListings       → remove feed-absent listings
  writeCycleStats         → per-search observability

fetchAndMatch is now a ~40-line orchestrator calling these in
sequence. No behavioral change — purely structural.

closes #1309
