+++ internal/scheduler/scheduler.go
@@ -731,508 +731,49 @@ func (s *Scheduler) fetchTargetedListings(ctx context.Context, searches []storag
 func (s *Scheduler) fetchAndMatch(ctx context.Context, searches []storage.Search, marketCache *scoring.MarketCache) (cycleStats, error) {
 	var stats cycleStats
 
+	// 1. Fetch raw listings from source.
 	fetchStart := time.Now()
-	raw, feedTokensByKey := s.fetchTargetedListings(ctx, searches, nil, s.targetedFetcher)
+	ms := s.fetchRawListings(ctx, searches)
 	fetchDuration := time.Since(fetchStart)
-	stats.listingsFetched = len(raw)
+	stats.listingsFetched = len(ms.raw)
 	s.observer.RecordFetch("yad2", fetchDuration, nil)
 
 	s.logger.Info("targeted listings fetched",
 		"scan", s.cycleCount,
 		"source", "yad2",
-		"listings", len(raw),
+		"listings", len(ms.raw),
 		"active_searches", len(searches),
 		"duration_ms", fetchDuration.Milliseconds(),
 	)
 
 	// 2. Catalog ingestion.
-	if s.catalogIngester != nil {
-		for _, l := range raw {
-			s.catalogIngester.Ingest(ctx, catalog.IngestEntry{
-				ManufacturerID:     l.ManufacturerID,
-				ManufacturerName:   l.Manufacturer,
-				ManufacturerNameHe: l.ManufacturerNameHe,
-				ModelID:            l.ModelID,
-				ModelName:          l.Model,
-				ModelNameHe:        l.ModelNameHe,
-			})
-		}
-	}
+	s.ingestCatalog(ctx, ms.raw)
 
-	// 3. Prefill from DB so listings with known km/city/image from prior
-	// cycles get that data before matching.
-	prefilled := false
+	// 3. Prefill from DB.
 	if s.stores.Listings != nil {
-		prefilled = s.prefillFromDB(ctx, raw)
+		ms.prefilled = s.prefillFromDB(ctx, ms.raw)
 	}
 
-	// 4. Build a per-search accumulator to collect results across listings.
-	type searchAccum struct {
-		search storage.Search
-		result searchResult
-		lang   locale.Lang
-	}
-	accums := make(map[int64]*searchAccum, len(searches))
+	// 4. Match listings to searches via percolator.
+	s.matchListingsToSearches(ctx, ms)
 
-	// Per-search cycle stats for observability.
-	type searchCycleCounter struct {
-		matched    int // percolator hits before dedup
-		kmFiltered int // held back by km filter
-	}
-	searchCounters := make(map[int64]*searchCycleCounter)
+	// 5. Publish enrichment requests.
+	s.publishEnrichRequests(ctx, ms)
 
-	// Deferred price drop candidates: we read prices during matching
-	// (to detect changes) but defer notification formatting until after
-	// enrichment so messages include enriched km/city data.
-	type priceDropCandidate struct {
-		rawIdx   int
-		searchID int64
-		oldPrice int
-	}
-	var priceDropCandidates []priceDropCandidate
+	// 6. Process price drops.
+	s.processPriceDrops(ctx, ms, marketCache)
 
-	// Price observations are staged here during matching and durably
-	// recorded only after the listing outcome is settled (see
-	// flushPendingPrices). persistedTokens / persistFailedTokens track
-	// per-token persist outcomes across all searches in this cycle.
-	pendingPrices := make(map[string]int)
-	persistedTokens := make(map[string]bool)
-	persistFailedTokens := make(map[string]bool)
+	// 7. Pipeline, persist, and deliver per search.
+	stats.notificationsSent = s.processAndDeliverPerSearch(ctx, ms, marketCache)
 
-	// 5. Percolator match: for each listing, find matching searches.
-	hiddenCache := make(map[int64]map[string]bool)
-	matchedIndices := make(map[int][]int64)
-	type priceResult struct {
-		oldPrice int
-		changed  bool
-		err      bool
-	}
-	priceCache := make(map[string]priceResult)
+	// 8. Flush staged prices.
+	s.flushPendingPrices(ctx, ms.pendingPrices, ms.persistedTokens, ms.persistFailedTokens)
 
-	for i := range raw {
-		matches := s.percolator.Match(raw[i])
-		if len(matches) == 0 {
-			continue
-		}
-		if _, seen := matchedIndices[i]; !seen {
-			ids := make([]int64, 0, len(matches))
-			for _, m := range matches {
-				ids = append(ids, m.SearchID)
-			}
-			matchedIndices[i] = ids
-		}
+	// 9. Drop stale listings.
+	droppedBySearch := s.dropStaleListings(ctx, searches, ms)
 
-		for _, m := range matches {
-			acc, ok := accums[m.SearchID]
-			if !ok {
-				lang := s.userLang(ctx, m.ChatID)
-				acc = &searchAccum{
-					search: m.Search,
-					lang:   lang,
-				}
-				accums[m.SearchID] = acc
-			}
-
-			ctr, ok := searchCounters[m.SearchID]
-			if !ok {
-				ctr = &searchCycleCounter{}
-				searchCounters[m.SearchID] = ctr
-			}
-			ctr.matched++
-
-			// Per-user hidden listing check (cached per chatID per cycle).
-			hidden, cached := hiddenCache[m.ChatID]
-			if !cached {
-				hidden = s.loadHiddenTokens(ctx, m.ChatID)
-				hiddenCache[m.ChatID] = hidden
-			}
-			if len(hidden) > 0 && hidden[raw[i].Token] {
-				continue
-			}
-
-			// Per-user, per-search dedup.
-			matchLog := s.logger.With("search_id", m.SearchID, "chat_id", m.ChatID, "search_name", m.SearchName)
-			isNew, ok := s.deduplicateListings(ctx, raw[i].Token, m.ChatID, m.SearchID, matchLog)
-			if !ok {
-				continue
-			}
-
-			// Read the last recorded price to detect changes; the durable
-			// write is staged and flushed only after the listing outcome
-			// is settled (flushPendingPrices). Cache per token so
-			// subsequent matches for the same listing reuse the read.
-			if s.stores.Prices != nil && raw[i].Price > 0 {
-				pr, cached := priceCache[raw[i].Token]
-				if !cached {
-					prevPrice, found, err := s.stores.Prices.PeekPrice(ctx, raw[i].Token)
-					if err != nil {
-						matchLog.ErrorContext(ctx, "failed to read last recorded listing price",
-							"token", raw[i].Token, "error", err.Error())
-						pr = priceResult{err: true}
-					} else {
-						pr = priceResult{oldPrice: prevPrice, changed: found && raw[i].Price != prevPrice}
-						if !found || pr.changed {
-							pendingPrices[raw[i].Token] = raw[i].Price
-						}
-					}
-					priceCache[raw[i].Token] = pr
-				}
-				if !pr.err && pr.changed && raw[i].Price < pr.oldPrice {
-					priceDropCandidates = append(priceDropCandidates, priceDropCandidate{
-						rawIdx: i, searchID: m.SearchID, oldPrice: pr.oldPrice,
-					})
-					continue
-				}
-			}
-
-			if !isNew {
-				continue
-			}
-
-			// Collect the raw listing for pipeline processing.
-			acc.result.newListings = append(acc.result.newListings, model.Listing{RawListing: raw[i]})
-		}
-	}
-
-	// 5a. Publish enrichment requests to the Redis stream for each matched
-	// listing that still needs enrichment data. The enricher worker
-	// processes these idempotently (skip if already enriched).
-	if s.enrichPublisher != nil && len(matchedIndices) > 0 {
-		published := 0
-		for idx, searchIDs := range matchedIndices {
-			l := raw[idx]
-			if l.Km <= 0 || l.City == "" || l.ImageURL == "" {
-				req := broker.EnrichRequest{
-					Token:      l.Token,
-					Priority:   1,
-					SearchIDs:  searchIDs,
-					Source:     "scheduler",
-					EnqueuedAt: time.Now().UTC().Format(time.RFC3339),
-				}
-				if err := s.enrichPublisher.PublishEnrich(ctx, req); err != nil {
-					s.logger.WarnContext(ctx, "failed to publish enrichment request to stream",
-						"token", l.Token, "car", l.Manufacturer+" "+l.Model, "error", err)
-				} else {
-					published++
-				}
-			}
-		}
-		if published > 0 {
-			s.logger.InfoContext(ctx, "published enrichment requests for matched listings",
-				"published", published, "total_matched", len(matchedIndices))
-		}
-	}
-
-	// 5b. Process deferred price drops.
-	for _, pd := range priceDropCandidates {
-		acc := accums[pd.searchID]
-		if acc == nil {
-			continue
-		}
-		l := raw[pd.rawIdx]
-		s.logger.InfoContext(ctx, "detected price drop for matched listing, preparing to notify user",
-			"token", l.Token, "old_price", pd.oldPrice, "new_price", l.Price,
-			"price_change", l.Price-pd.oldPrice,
-			"manufacturer", l.Manufacturer, "model", l.Model, "year", l.Year,
-		)
-		listing := model.Listing{RawListing: l, SearchName: acc.search.Name}
-		fp := scoring.FitnessParams{
-			Price: l.Price, Km: l.Km, Hand: l.Hand, Year: l.Year,
-			EngineVolume: l.EngineVolume, PriceMax: acc.search.PriceMax,
-			MaxKm: acc.search.MaxKm, MaxHand: acc.search.MaxHand,
-			YearMin: acc.search.YearMin, YearMax: acc.search.YearMax,
-			EngineMinCC: acc.search.EngineMinCC,
-		}
-		if marketCache != nil {
-			if median, medKm, _, ok := marketCache.Lookup(l.Manufacturer, l.Model, l.Year); ok {
-				fp.MedianPrice = median
-				fp.MedianKm = medKm
-			}
-		}
-		listing.FitnessScore = scoring.FitnessScore(fp)
-		acc.result.priceDropMessages = append(acc.result.priceDropMessages, notifier.FormatPriceDrop(listing, pd.oldPrice, acc.lang))
-		if s.stores.Listings != nil {
-			rec := storage.ListingRecord{
-				Token: l.Token, ChatID: acc.search.ChatID, SearchID: acc.search.ID, SearchName: acc.search.Name,
-				Manufacturer: l.Manufacturer, Model: l.Model, SubModel: l.SubModel,
-				SubModelID: l.SubModelID,
-				Year:       l.Year, Price: l.Price, Km: l.Km, Hand: l.Hand,
-				City: l.City, PageLink: l.PageLink, ImageURL: l.ImageURL,
-				EngineVolume: l.EngineVolume, HorsePower: l.HorsePower,
-				EngineType: l.EngineType, GearBox: l.GearBox, Description: l.Description,
-				IsCommercial: l.Commercial,
-				FitnessScore: &listing.FitnessScore, FirstSeenAt: time.Now(),
-			}
-			if !l.CreatedAt.IsZero() {
-				t := l.CreatedAt
-				rec.PostedAt = &t
-			}
-			if marketCache != nil {
-				if median, medKm, cohort, ok := marketCache.Lookup(l.Manufacturer, l.Model, l.Year); ok {
-					ds := scoring.ScoreWithKm(l.Price, l.Km, median, medKm)
-					rec.MedianPrice = &median
-					rec.CohortSize = &cohort
-					rec.DealScore = &ds
-				}
-			}
-			if err := s.stores.Listings.SaveListing(ctx, rec); err != nil {
-				s.logger.ErrorContext(ctx, "failed to persist price-drop listing to history",
-					"token", l.Token, "error", err.Error())
-			}
-		}
-	}
-
-	// Build a token→index map so the pipeline reads enriched data from raw.
-	rawByToken := make(map[string]int, len(raw))
-	for i := range raw {
-		rawByToken[raw[i].Token] = i
-	}
-
-	// 6. Run pipeline and deliver results per search.
-	for _, acc := range accums {
-		searchCtx := cwlog.WithSearch(ctx, acc.search.ID, acc.search.ChatID)
-
-		// Convert collected raw listings through the pipeline, using the
-		// enriched version from raw (not the stale copy in newListings).
-		if len(acc.result.newListings) > 0 {
-			rawForPipeline := make([]model.RawListing, len(acc.result.newListings))
-			for i, l := range acc.result.newListings {
-				if idx, ok := rawByToken[l.Token]; ok {
-					rawForPipeline[i] = raw[idx]
-				} else {
-					rawForPipeline[i] = l.RawListing
-				}
-			}
-
-			// Process ALL matched listings through the pipeline so they
-			// get persisted to listing_history. The enricher needs the
-			// rows to exist before it can backfill km/city data.
-			if len(rawForPipeline) == 0 {
-				acc.result.newListings = nil
-			} else {
-				params := ProcessParamsFromSearch(acc.search, marketCache)
-				params.SkipPrefill = prefilled
-				pr := s.pipeline.Process(searchCtx, rawForPipeline, params)
-				acc.result.newListings = pr.Listings
-				acc.result.listingRecords = pr.Records
-			}
-
-			// Post-enrichment MaxKm filter: suppress notifications for
-			// listings whose km exceeds the search cap or is still
-			// unknown (pending enrichment). Listings are still persisted
-			// above so the enricher can backfill km data.
-			if acc.search.MaxKm > 0 && len(acc.result.newListings) > 0 {
-				var deliverable []model.Listing
-				for _, l := range acc.result.newListings {
-					if l.Km <= 0 || l.Km > acc.search.MaxKm {
-						reason := "km_exceeded"
-						if l.Km <= 0 {
-							reason = "km_pending_enrichment"
-						}
-						if relErr := s.stores.Dedup.ReleaseClaim(ctx, l.Token, acc.search.ChatID, acc.search.ID); relErr != nil {
-							s.logger.ErrorContext(searchCtx,
-								"failed to release dedup claim for km-filtered listing",
-								"token", l.Token, "reason", reason, "error", relErr.Error())
-						}
-						if ctr := searchCounters[acc.search.ID]; ctr != nil {
-							ctr.kmFiltered++
-						}
-						continue
-					}
-					deliverable = append(deliverable, l)
-				}
-				if len(deliverable) < len(acc.result.newListings) {
-					s.logger.InfoContext(searchCtx,
-						"filtered listings by km limit after enrichment",
-						"search_name", acc.search.Name,
-						"chat_id", acc.search.ChatID,
-						"before", len(acc.result.newListings),
-						"after", len(deliverable),
-						"max_km", acc.search.MaxKm,
-					)
-					acc.result.newListings = deliverable
-				}
-			}
-		}
-
-		// Persist listings. Per-token outcomes feed flushPendingPrices: a
-		// staged price is only written durably once its listing made it
-		// into history somewhere.
-		searchLog := s.logger.With("chat_id", acc.search.ChatID, "search_name", acc.search.Name)
-		persistOK := true
-		if s.stores.Listings != nil && len(acc.result.listingRecords) > 0 {
-			if err := s.persistListings(searchCtx, acc.result.listingRecords, searchLog); err != nil {
-				persistOK = false
-				for _, rec := range acc.result.listingRecords {
-					persistFailedTokens[rec.Token] = true
-				}
-			} else {
-				s.invalidateMarketCache()
-				for _, rec := range acc.result.listingRecords {
-					persistedTokens[rec.Token] = true
-				}
-			}
-		}
-		if !persistOK {
-			acc.result.newListings = nil
-			acc.result.listingRecords = nil
-		}
-
-		stats.newListings += len(acc.result.newListings)
-
-		// Enrich grace window: defer notification for unenriched listings
-		// that were just seen, giving the enricher time to fill in km/city/image.
-		graceSec := s.cfg.Polling.EnrichGraceSeconds
-		if graceSec <= 0 {
-			graceSec = 60
-		}
-		if graceSec > 0 && len(acc.result.newListings) > 0 {
-			var ready []model.Listing
-			deferred := 0
-			for _, l := range acc.result.newListings {
-				unenriched := l.Km <= 0 || l.City == "" || l.ImageURL == ""
-				justSeen := !l.CreatedAt.IsZero() && time.Since(l.CreatedAt) < time.Duration(graceSec)*time.Second
-				if unenriched && justSeen {
-					deferred++
-					continue
-				}
-				ready = append(ready, l)
-			}
-			if deferred > 0 {
-				searchLog.InfoContext(searchCtx, "deferred unenriched listings for enrich grace window",
-					"deferred", deferred, "ready", len(ready), "grace_seconds", graceSec)
-			}
-			acc.result.newListings = ready
-		}
-
-		// Deliver notifications.
-		if len(acc.result.newListings) > 0 || len(acc.result.priceDropMessages) > 0 {
-			searchLog.InfoContext(searchCtx, "search matched listings, preparing delivery",
-				"new", len(acc.result.newListings),
-				"price_drops", len(acc.result.priceDropMessages),
-			)
-		}
-		delivered := s.deliverResults(searchCtx, acc.search, acc.lang, acc.result, searchLog)
-		if delivered {
-			stats.notificationsSent++
-		}
-	}
-
-	// Durably record staged price observations now that listing outcomes
-	// are settled.
-	s.flushPendingPrices(ctx, pendingPrices, persistedTokens, persistFailedTokens)
-
-	// Drop stale listings that no longer appear in the source feed.
-	droppedBySearch := make(map[int64]int64)
-	if s.stores.Listings != nil && len(feedTokensByKey) > 0 {
-		for _, search := range searches {
-			params := model.SourceParamsFromSearch(&search)
-			key := fetcher.CacheKeyFor(params)
-			feedTokens, ok := feedTokensByKey[key]
-			if !ok {
-				continue
-			}
-			dropped, err := s.stores.Listings.DeleteStaleListings(ctx, search.ChatID, search.ID, feedTokens)
-			if err != nil {
-				s.logger.WarnContext(ctx, "failed to drop stale listings",
-					"search_id", search.ID, "chat_id", search.ChatID, "error", err)
-				continue
-			}
-			if dropped > 0 {
-				droppedBySearch[search.ID] = dropped
-				s.logger.InfoContext(ctx, "dropped stale listings",
-					"search_id", search.ID, "search_name", search.Name,
-					"chat_id", search.ChatID, "dropped", dropped)
-			}
-		}
-	}
-
-	// Write per-search cycle stats for dashboard observability.
-	if s.stores.SearchCycleStats != nil && len(searches) > 0 {
-		rejections := s.percolator.CountRejections(raw)
-		now := time.Now()
-
-		cycleStats := make([]storage.SearchCycleStats, 0, len(searches))
-		for _, search := range searches {
-			acc := accums[search.ID]
-			ctr := searchCounters[search.ID]
-			matched := 0
-			kmFiltered := 0
-			if ctr != nil {
-				matched = ctr.matched
-				kmFiltered = ctr.kmFiltered
-			}
-
-			rej := rejections[search.ID]
-
-			var newListingCount, priceDropCount int
-			var scoreMin, scoreMax, scoreAvg *float64
-			var priceMin, priceMax *int
-			if acc != nil && len(acc.result.newListings) > 0 {
-				newListingCount = len(acc.result.newListings)
-				priceDropCount = len(acc.result.priceDropMessages)
-				mn, mx, sum := acc.result.newListings[0].FitnessScore, acc.result.newListings[0].FitnessScore, 0.0
-				pMin, pMax := acc.result.newListings[0].Price, acc.result.newListings[0].Price
-				for _, l := range acc.result.newListings {
-					if l.FitnessScore < mn {
-						mn = l.FitnessScore
-					}
-					if l.FitnessScore > mx {
-						mx = l.FitnessScore
-					}
-					sum += l.FitnessScore
-					if l.Price > 0 {
-						if pMin == 0 || l.Price < pMin {
-							pMin = l.Price
-						}
-						if l.Price > pMax {
-							pMax = l.Price
-						}
-					}
-				}
-				avg := sum / float64(len(acc.result.newListings))
-				scoreMin = &mn
-				scoreMax = &mx
-				scoreAvg = &avg
-				if pMin > 0 {
-					priceMin = &pMin
-					priceMax = &pMax
-				}
-			} else if acc != nil {
-				priceDropCount = len(acc.result.priceDropMessages)
-			}
-
-			cycleStats = append(cycleStats, storage.SearchCycleStats{
-				SearchID:    search.ID,
-				ChatID:      search.ChatID,
-				SearchName:  search.Name,
-				CycleAt:     now,
-				FeedSize:    matched + rej[percolator.RejectYearOut] + rej[percolator.RejectPriceOut] + rej[percolator.RejectKmOver] + rej[percolator.RejectHandOver] + rej[percolator.RejectEngineCC] + rej[percolator.RejectSeller] + rej[percolator.RejectOtherFilter],
-				Matched:     matched,
-				NewListings: newListingCount,
-				KmFiltered:  kmFiltered,
-				Delivered:   newListingCount,
-				PriceDrops:  priceDropCount,
-				WrongModel:  rej[percolator.RejectWrongModel],
-				YearOut:     rej[percolator.RejectYearOut],
-				PriceOut:    rej[percolator.RejectPriceOut],
-				KmOver:      rej[percolator.RejectKmOver],
-				HandOver:    rej[percolator.RejectHandOver],
-				EngineCC:    rej[percolator.RejectEngineCC],
-				Seller:      rej[percolator.RejectSeller],
-				OtherFilter: rej[percolator.RejectOtherFilter],
-				Dropped:     int(droppedBySearch[search.ID]),
-				ScoreMin:    scoreMin,
-				ScoreMax:    scoreMax,
-				ScoreAvg:    scoreAvg,
-				PriceMin:    priceMin,
-				PriceMax:    priceMax,
-			})
-		}
-		if err := s.stores.SearchCycleStats.UpsertSearchCycleStats(ctx, cycleStats); err != nil {
-			s.logger.Error("failed to write per-search cycle stats", "error", err)
-		}
-	}
+	// 10. Write cycle stats.
+	s.writeCycleStats(ctx, searches, ms, ms.raw, droppedBySearch)
 
 	return stats, nil
 }
